% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_from_fukan}
\alias{read_from_fukan}
\title{Read files generated in Fukan System.}
\usage{
read_from_fukan(path_to_file)
}
\arguments{
\item{path_to_file}{The full path of the file you want to open}
}
\description{
This function allows you to read report analysis from Fukan System
}
\examples{
read_from_fukan(file.choose())
}
\seealso{
\code{\link{write_for_fukan}}
}
\keyword{Fukan}
\keyword{System}
