% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_from_others_tsv}
\alias{read_from_others_tsv}
\title{Read .TSV data frames.}
\usage{
read_from_others_tsv(path_to_file)
}
\arguments{
\item{path_to_file}{The full path of the file you want to open}
}
\description{
This function allows you to read .TSV files.
}
\details{
Files are expected to have header. Contents are forced to strings.
}
\examples{
read_from_others_tsv(file.choose())
}
\seealso{
\code{\link{read_from_others_csv}}
}
