% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_from_ti_fast}
\alias{read_from_ti_fast}
\title{Fast reading of patent report data from Thomson Innovation.}
\usage{
read_from_ti_fast(path_to_file, skip = 1)
}
\arguments{
\item{path_to_file}{The full path of the file you want to open}

\item{skip}{Defaults to 1. The number of rows to skip before starting readind the data}
}
\description{
This function allows you to read .CSV reports from Thomson Innovation.
}
\details{
Encoding problem raises when Thomson Innovation interface is in Japanese, using English interface is advised.
}
\examples{
read_from_ti_fast(file.choose())
}
\seealso{
\code{\link{read_from_ti}}
}
\keyword{Innovation}
\keyword{Thomson}
