% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_from_wos}
\alias{read_from_wos}
\title{Read biblographic data from Web of Science.}
\usage{
read_from_wos(path_to_file)
}
\arguments{
\item{path_to_file}{The full path of the file you want to open}
}
\description{
This function allows you to read WIN UTF 8, TSV files from the Web of Science.
}
\examples{
read_from_wos(file.choose())
}
\keyword{Science}
\keyword{Web}
\keyword{of}
