% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_for_fukan}
\alias{write_for_fukan}
\title{Writes data sets to be used in Fukan System}
\usage{
write_for_fukan(data_file, file_name)
}
\arguments{
\item{data_file}{The data frame to use}

\item{file_name}{The name of the file in the format "*.tsv"}
}
\description{
This function writes files that can be used again in Fukan system to create network analysis.
}
\examples{
write_for_fukan(my_data, 'my_data.tsv')
}
\seealso{
\code{\link{read_from_fukan}}
}
\keyword{Fukan}
\keyword{System}
